田梓霖修改xml分页方法

+++ food_map/src/main/java/com/gok/food_map/product/dto/ProductsGetListDto.java
@@ -9,10 +9,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ProductsGetListDto {
     //当前页
-    private Integer current;
+    private Long current;
 
     //页大小
-    private Integer size;
+    private Long size;
 
     //编号
     private String spuId;
